backend/requirements/requirements.txt
Flask
Flask-SQLAlchemy
ariadne
requests
# psycopg2 #prob need binary
psycopg2-binary
flask-graphql
#>=2.0.1
boto3==1.18.0
selenium
bs4
graphql-core
Flask-Migrate==3.1.0
flask-cors==3.0.10


